feat: Add uniform contracts system foundation
Core contracts system with uniform parameters across CLI/MCP/HTTP interfaces.
Includes comprehensive testing (9/9 tests pass), service layer, and adapters.

Note: Contains work-in-progress CLI implementation with complexity to be
addressed in next commit. Foundation is complete and working.

+++ server/src/cli/handlers/analysis_handlers.rs
@@ -446,7 +446,7 @@ pub async fn route_analyze_command(cmd: AnalyzeCommands) -> Result<()> {
         } => {
             super::defect_prediction_handler::handle_analyze_defect_prediction(
                 project_path,
-                confidence_threshold as f32,
+                confidence_threshold,
                 min_lines,
                 include_low_confidence,
                 format,
